backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
PyJWT==2.8.0
bcrypt==4.0.1
pymongo==4.6.0